+++ Basil/include/Core/Input.h
@@ -8,19 +8,21 @@
 #include <utility>
 
 #include "Core/Core.h"
+#include "Core/KeyCodes.h"
+#include "Core/MouseCodes.h"
 
 namespace Basil
 {
 	class Input
 	{
 		public:
-			static bool isKeyPressed(int keycode);
-			static bool isMouseButtonPressed(int button);
+			static bool isKeyPressed(KeyCode keyCode);
+			static bool isMouseButtonPressed(MouseCode mouseCode);
 			static float getMouseX();
 			static float getMouseY();
 		protected:
-			virtual bool isKeyPressedImpl(int keycode) = 0;
-			virtual bool isMouseButtonPressedImpl(int button) = 0;
+			virtual bool isKeyPressedImpl(KeyCode keyCode) = 0;
+			virtual bool isMouseButtonPressedImpl(MouseCode mouseCode) = 0;
 			virtual std::pair<float, float> getMousePositionImpl() = 0;
 			virtual float getMouseXImpl() = 0;
 			virtual float getMouseYImpl() = 0;