+++ Basil/src/Core/Input.cpp
@@ -5,15 +5,15 @@ namespace Basil
 {
 	// All functions call the implementation (which is platform-specific)
 	// Return if key is pressed
-	bool Input::isKeyPressed(int keycode)
+	bool Input::isKeyPressed(KeyCode keycode)
 	{
 		return instance->isKeyPressedImpl(keycode);
 	}
 
 	// Return if mouse button is pressed
-	bool Input::isMouseButtonPressed(int button)
+	bool Input::isMouseButtonPressed(MouseCode mouseCode)
 	{
-		return instance->isMouseButtonPressedImpl(button);
+		return instance->isMouseButtonPressedImpl(mouseCode);
 	}
 
 	// Return mouse X position